+++ CHANGELOG.md
@@ -17,6 +17,11 @@
     Click to see more.
   </summary>
 
+## :bug: Bug Fix
+
+- `tiptap-editor`
+  - [#4522](https://github.com/wix/ricos/pull/4522) editor missing padding in mobile + remove image sizing handles on mobile
+
 </details>
 
 ## 8.72.45 (Sep 1, 2022)